Meter WIP and animal and egg meshes

+++ Cowotchi/src/Core/Meters/MeterObservable.cs
@@ -4,29 +4,29 @@ using System;
 public partial class MeterObservable : Node
 {
     [Signal]
-    public delegate void UpdateHeartMeterValueEventHandler(float value);
-    public void EmitUpdateHeartMeterValue(float value)
+    public delegate void UpdateHeartMeterValueEventHandler(int value);
+    public void EmitUpdateHeartMeterValue(int value)
     {
         EmitSignal(SignalName.UpdateHeartMeterValue, value);
     }
 
     [Signal]
-    public delegate void UpdateHeartMeterMaxEventHandler(float max);
-    public void EmitUpdateHeartMeterMax(float max)
+    public delegate void UpdateHeartMeterMaxEventHandler(int max);
+    public void EmitUpdateHeartMeterMax(int max)
     {
         EmitSignal(SignalName.UpdateHeartMeterMax, max);
     }
 
     [Signal]
-    public delegate void UpdateHungerMeterValueEventHandler(float value);
-    public void EmitUpdateHungerMeterValue(float value)
+    public delegate void UpdateHungerMeterValueEventHandler(int value);
+    public void EmitUpdateHungerMeterValue(int value)
     {
         EmitSignal(SignalName.UpdateHungerMeterValue, value);
     }
 
     [Signal]
-    public delegate void UpdateHungerMeterMaxEventHandler(float max);
-    public void EmitUpdateHungerMeterMax(float max)
+    public delegate void UpdateHungerMeterMaxEventHandler(int max);
+    public void EmitUpdateHungerMeterMax(int max)
     {
         EmitSignal(SignalName.UpdateHungerMeterMax, max);
     }
